Move components from components root
