+++ src/pages/AvailabilityPageMultiple.js
