+++ src/pages/BookingAvailabilityModal.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { BookingAvailabilityForm } from '@/components/BookingAvailabilityForm';
+import { BookingAvailabilityForm } from '@/pages/BookingAvailabilityForm';
 import { Modal, ModalSizes, ModalVariants } from '@/ui/Modal';
 import { Spinner } from '@/ui/Spinner';
 import { Stack } from '@/ui/Stack';
