+++ tsconfig.json
@@ -9,7 +9,6 @@
       "@/pages/*": ["components/pages/*"],
       "@/utils/*": ["utils/*"],
       "@/constants/*": ["constants/*"],
-      "@/components/*": ["components/*"],
       "@/test/*": ["test/*"],
       "@/i18n/*": ["i18n/*"],
       "@/layouts/*": ["layouts/*"],
